[Structure] Moved files into new structure. Models contains in a models package and utility functionality in a utils package
