+++ src/ronald_brain/models/markov_chain.py
@@ -1,5 +1,7 @@
 from collections import defaultdict
-from .constants import *
+
+from ronald_brain.util.constants import *
+
 
 class MarkovChain():
     def __init__(self, n_tuples, order):
